@@ -1,6 +1,85 @@
+[build-system]
+requires = ["setuptools>=65.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "cods"
+version = "0.3"
+description = "Conformal Object Detection & Segmentation"
+readme = "README.md"
+requires-python = ">=3.10"
+license = {text = "MIT"}
+authors = [
+    {name = "Léo Andéol", email = "leo@andeol.eu"},
+    {name = "Luca Mossina"},
+]
+keywords = ["conformal-prediction", "object-detection", "segmentation", "computer-vision", "pytorch"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+]
+
+# Core dependencies - pinned for stability
+dependencies = [
+    "torch>=2.0.0",
+    "torchvision>=0.15.0",
+    "numpy>=2.2.6",
+    "pillow>=9.0.0",
+    "tqdm>=4.60.0",
+    "numba>=0.56.0",
+    "glob2>=0.7",
+    "matplotlib>=3.5.0",
+    "scikit-learn>=1.0.0",
+    "scikit-optimize>=0.9.0",
+    "timm>=0.9.0",               # PyTorch Image Models
+    "pycocotools>=2.0.6",
+    "python-dotenv>=0.19.0",
+    "ultralytics>=8.0.0",        # YOLO models
+]
+
+[project.optional-dependencies]
+# Development tools
+dev = [
+    "ruff>=0.4.0",
+    "mypy>=1.0.0",
+    "pre-commit>=3.0.0",
+    "pytest>=7.0.0",
+    "pytest-cov>=4.0.0",
+    "pytest-xdist>=3.0.0",
+    "pytest-timeout>=2.0.0",
+    "ipykernel>=6.0.0",
+    "jupyter>=1.0.0",
+    "notebook>=6.4.0",
+    "wandb>=0.15.0",             # Weights & Biases for experiment tracking
+    "line-profiler>=4.0.0",      # Performance profiling
+    "sphinx>=8.1.3",
+    "sphinx-rtd-theme>=3.0.2",
+    "sphinx-autodoc-typehints>=3.0.1",
+]
+
+# All dependencies
+all = ["cods[dev]"]
+
+[project.urls]
+Homepage = "https://github.com/leoandeol/cods/"
+Repository = "https://github.com/leoandeol/cods/"
+Issues = "https://github.com/leoandeol/cods/issues"
+
+# Tell setuptools where to find packages
+[tool.setuptools.packages.find]
+include = ["cods*"]
+namespaces = true
+
+# Ruff configuration
 [tool.ruff]
-target-version = "py38"
-# Exclude a variety of commonly ignored directories.
+target-version = "py310"
+line-length = 100
+indent-width = 4
+
 exclude = [
     ".bzr",
     ".direnv",
