@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import logging
-from typing import List, Optional, Tuple, Union
+from types import MappingProxyType
 
 import torch
 
@@ -52,7 +52,9 @@ from cods.od.utils import (
 
 logger = logging.getLogger("cods")
 # FORMAT = "[%(filename)s:%(lineno)s - %(funcName)20s() ] %(message)s"
-FORMAT = "[%(asctime)s:%(levelname)s:%(filename)s:%(module)s:%(lineno)s - %(funcName)s ] %(message)s"
+FORMAT = (
+    "[%(asctime)s:%(levelname)s:%(filename)s:%(module)s:%(lineno)s - %(funcName)s ] %(message)s"
+)
 logging.basicConfig(format=FORMAT, datefmt="%Y-%m-%d %H:%M:%S")
 
 """
@@ -82,28 +84,34 @@ class LocalizationConformalizer(Conformalizer):
 
     """
 
-    BACKENDS = ["auto", "cp", "crc"]
+    BACKENDS = ("auto", "cp", "crc")
 
-    accepted_methods = {
-        "min-hausdorff-additive": MinAdditiveSignedAssymetricHausdorffNCScore,
-        "min-hausdorff-multiplicative": MinMultiplicativeSignedAssymetricHausdorffNCScore,
-        "union-hausdorff-additive": UnionAdditiveSignedAssymetricHausdorffNCScore,
-        "union-hausdorff-multiplicative": UnionMultiplicativeSignedAssymetricHausdorffNCScore,
-    }
+    accepted_methods = MappingProxyType(
+        {
+            "min-hausdorff-additive": MinAdditiveSignedAssymetricHausdorffNCScore,
+            "min-hausdorff-multiplicative": MinMultiplicativeSignedAssymetricHausdorffNCScore,
+            "union-hausdorff-additive": UnionAdditiveSignedAssymetricHausdorffNCScore,
+            "union-hausdorff-multiplicative": UnionMultiplicativeSignedAssymetricHausdorffNCScore,
+        }
+    )
 
-    PREDICTION_SETS = ["additive", "multiplicative", "adaptive"]
-    LOSSES = {
-        "pixelwise": PixelWiseRecallLoss,
-        "boxwise": BoxWiseRecallLoss,
-        "thresholded": ThresholdedRecallLoss,
-        "boxwise-precision": BoxWisePrecisionLoss,
-        "boxwise-iou": BoxWiseIoULoss,
-    }
-    OPTIMIZERS = {
-        "binary_search": BinarySearchOptimizer,
-        "gaussian_process": GaussianProcessOptimizer,
-    }
-    GUARANTEE_LEVELS = ["image", "object"]
+    PREDICTION_SETS = ("additive", "multiplicative", "adaptive")
+    LOSSES = MappingProxyType(
+        {
+            "pixelwise": PixelWiseRecallLoss,
+            "boxwise": BoxWiseRecallLoss,
+            "thresholded": ThresholdedRecallLoss,
+            "boxwise-precision": BoxWisePrecisionLoss,
+            "boxwise-iou": BoxWiseIoULoss,
+        }
+    )
+    OPTIMIZERS = MappingProxyType(
+        {
+            "binary_search": BinarySearchOptimizer,
+            "gaussian_process": GaussianProcessOptimizer,
+        }
+    )
+    GUARANTEE_LEVELS = ("image", "object")
 
     def __init__(
         self,
@@ -112,7 +120,7 @@ class LocalizationConformalizer(Conformalizer):
         guarantee_level: str,
         matching_function: str,
         number_of_margins: int = 1,  # where to compute 1, 2 or 4 margins with bonferroni corrections
-        optimizer: Optional[str | Optimizer] = None,
+        optimizer: str | Optimizer | None = None,
         backend: str = "auto",
         device: str = "cpu",
     ):
@@ -208,12 +216,8 @@ class LocalizationConformalizer(Conformalizer):
         predictions: ODPredictions,
         alpha: float,
         steps: int = 13,
-        bounds: List[float] = [
-            0,
-            1000,
-        ],  # TODO: currently ignored by new optimizer
+        overload_confidence_threshold: float | None = None,
         verbose: bool = True,
-        overload_confidence_threshold: Optional[float] = None,
     ) -> float:
         """Calibrate the conformalizer.
 
@@ -222,7 +226,6 @@ class LocalizationConformalizer(Conformalizer):
         - predictions (ODPredictions): The object detection predictions.
         - alpha (float): The significance level.
         - steps (int): The number of steps for optimization.
-        - bounds (List[float]): The bounds for optimization.
         - verbose (bool): Whether to print the optimization progress.
         - confidence_threshold (float): The threshold for objectness confidence.
 
@@ -233,22 +236,6 @@ class LocalizationConformalizer(Conformalizer):
         """
         if self.lambda_localization is not None:
             logger.info("Replacing previously computed λ")
-        if overload_confidence_threshold is None:
-            if predictions.confidence_threshold is None:
-                raise ValueError(
-                    "confidence_threshold must be set in the predictions or in the conformalizer",
-                )
-            confidence_threshold = predictions.confidence_threshold
-            if isinstance(confidence_threshold, torch.Tensor):
-                confidence_threshold = confidence_threshold.item()
-            logger.info(
-                f"Using predictions' confidence threshold: {confidence_threshold:.4f}",
-            )
-        else:
-            logger.info(
-                f"Using overload confidence threshold: {overload_confidence_threshold:.4f}",
-            )
-            confidence_threshold = overload_confidence_threshold
 
         def build_predictions(
             matched_pred_boxes_i,
@@ -272,15 +259,15 @@ class LocalizationConformalizer(Conformalizer):
         lambda_localization = self.optimizer2.optimize(
             predictions,
             build_predictions,
+            overload_confidence_threshold=overload_confidence_threshold,
             loss=self.loss,
             matching_function=self.matching_function,
             alpha=alpha,
             device=self.device,
             B=1,
-            bounds=[0, 1000]
-            if self.prediction_set == "additive"
-            else [0, 100],
-            steps=13,
+            lower_bound=0,
+            upper_bound=1000 if self.prediction_set == "additive" else 100,
+            steps=steps,
             epsilon=1e-9,
             verbose=verbose,
         )
@@ -296,9 +283,9 @@ class LocalizationConformalizer(Conformalizer):
     def conformalize(
         self,
         predictions: ODPredictions,
-        parameters: Optional[ODParameters] = None,
+        parameters: ODParameters | None = None,
         verbose: bool = True,
-    ) -> List[torch.Tensor]:
+    ) -> list[torch.Tensor]:
         """Conformalizes the predictions using the specified lambda values for localization.
 
         Args:
@@ -322,10 +309,7 @@ class LocalizationConformalizer(Conformalizer):
             raise ValueError(
                 "Conformalizer must be calibrated, or parameters provided, before conformalizing.",
             )
-        if (
-            parameters is not None
-            and parameters.lambda_localization is not None
-        ):
+        if parameters is not None and parameters.lambda_localization is not None:
             if verbose:
                 logger.info("Using λ for localization from parameters")
             lambda_localization = parameters.lambda_localization
@@ -357,19 +341,21 @@ class LocalizationConformalizer(Conformalizer):
 class ConfidenceConformalizer(Conformalizer):
     """ """
 
-    ACCEPTED_LOSSES = {
-        "box_count_threshold": BoxCountThresholdConfidenceLoss,
-        "box_count_recall": BoxCountRecallConfidenceLoss,
-        "box_thresholded_distance": ThresholdedBoxDistanceConfidenceLoss,
-        "box_count_twosided_recall": BoxCountTwosidedConfidenceLoss,
-    }
+    ACCEPTED_LOSSES = MappingProxyType(
+        {
+            "box_count_threshold": BoxCountThresholdConfidenceLoss,
+            "box_count_recall": BoxCountRecallConfidenceLoss,
+            "box_thresholded_distance": ThresholdedBoxDistanceConfidenceLoss,
+            "box_count_twosided_recall": BoxCountTwosidedConfidenceLoss,
+        }
+    )
 
     def __init__(
         self,
         guarantee_level: str,
         matching_function: str,
         loss: str = "box_count_threshold",
-        other_losses: Optional[List] = None,
+        other_losses: list | None = None,
         optimizer: str = "binary_search",
         device="cpu",
     ):
@@ -411,10 +397,12 @@ class ConfidenceConformalizer(Conformalizer):
         predictions: ODPredictions,
         alpha: float = 0.1,
         steps: int = 13,
-        bounds: List[float] = [0, 1],
+        bounds: list[float] | None = None,
         verbose: bool = True,
-    ) -> Tuple[float, float]:
+    ) -> tuple[float, float]:
         """ """
+        if bounds is None:
+            bounds = [0, 1]
         if self.lambda_plus is not None:
             logger.info("Replacing previously computed λ")
 
@@ -429,7 +417,6 @@ class ConfidenceConformalizer(Conformalizer):
             alpha,
             self.device,
             B=1,
-            bounds=[0, 1],
             verbose=False,
         )
         logger.debug("Optimizing for lambda_minus")
@@ -443,7 +430,6 @@ class ConfidenceConformalizer(Conformalizer):
             alpha,
             self.device,
             B=0,
-            bounds=[0, 1],
             verbose=False,
         )
         self.lambda_plus = lambda_plus
@@ -478,15 +464,19 @@ class ConfidenceConformalizer(Conformalizer):
 class ODClassificationConformalizer(ClassificationConformalizer):
     """ """
 
-    BACKENDS = ["auto", "cp", "crc"]
-    GUARANTEE_LEVELS = ["image", "object"]
-    OPTIMIZERS = {
-        "binary_search": BinarySearchOptimizer,
-        "gaussian_process": GaussianProcessOptimizer,
-    }
-    LOSSES = {
-        "binary": ODBinaryClassificationLoss,
-    }
+    BACKENDS = ("auto", "cp", "crc")
+    GUARANTEE_LEVELS = ("image", "object")
+    OPTIMIZERS = MappingProxyType(
+        {
+            "binary_search": BinarySearchOptimizer,
+            "gaussian_process": GaussianProcessOptimizer,
+        }
+    )
+    LOSSES = MappingProxyType(
+        {
+            "binary": ODBinaryClassificationLoss,
+        }
+    )
 
     def __init__(
         self,
@@ -554,27 +544,15 @@ class ODClassificationConformalizer(ClassificationConformalizer):
         self,
         predictions: ODPredictions,
         alpha: float,
-        bounds: List[float] = [0, 1],
+        bounds: list[float] | None = None,
         steps: int = 40,
         verbose: bool = True,
-        overload_confidence_threshold: Optional[float] = None,
+        overload_confidence_threshold: float | None = None,
     ) -> torch.Tensor:
-        if overload_confidence_threshold is None:
-            if predictions.confidence_threshold is None:
-                raise ValueError(
-                    "confidence_threshold must be set in the predictions or in the conformalizer",
-                )
-            confidence_threshold = predictions.confidence_threshold
-            if isinstance(confidence_threshold, torch.Tensor):
-                confidence_threshold = confidence_threshold.item()
-            logger.info(
-                f"Using predictions' confidence threshold: {confidence_threshold:.4f}",
-            )
-        else:
-            logger.info(
-                f"Using overload confidence threshold: {overload_confidence_threshold:.4f}",
-            )
-            confidence_threshold = overload_confidence_threshold
+        if bounds is None:
+            bounds = [0, 1]
+        if self.lambda_classification is not None:
+            logger.info("Replacing previously computed λ")
 
         logger.warning(
             "Currently considering that there is only one matching prediction to each true box for classification pruposes. To add later how to aggregate if multiple preidctions matched.",
@@ -595,7 +573,7 @@ class ODClassificationConformalizer(ClassificationConformalizer):
             # TODO(leo): filter for confidence here!
             def get_conf_cls():
                 conf_cls_i = []
-                for j, pred_cls_i_j in enumerate(matched_pred_cls_i):
+                for _j, pred_cls_i_j in enumerate(matched_pred_cls_i):
                     conf_cls_i_j = self._score_function.get_set(
                         pred_cls=pred_cls_i_j,
                         quantile=lbd,
@@ -615,12 +593,14 @@ class ODClassificationConformalizer(ClassificationConformalizer):
         lambda_classification = self.optimizer2.optimize(
             predictions,
             build_predictions,
+            overload_confidence_threshold=overload_confidence_threshold,
             loss=self.loss,
             matching_function=self.matching_function,
             alpha=alpha,
             device=self.device,
             B=1,
-            bounds=[0, 1],
+            lower_bound=0,
+            upper_bound=1,
             steps=25,
             epsilon=1e-10,
             verbose=verbose,
@@ -637,14 +617,14 @@ class ODClassificationConformalizer(ClassificationConformalizer):
         self,
         predictions: ODPredictions,
         verbose: bool = True,
-    ) -> List:
+    ) -> list:
         # TODO: add od parameters to function signature
         # NO MATCHING HERE
         def get_conf_cls():
             conf_cls = []
-            for i, pred_cls_i in enumerate(predictions.pred_cls):
+            for _i, pred_cls_i in enumerate(predictions.pred_cls):
                 conf_cls_i = []
-                for j, pred_cls_i_j in enumerate(pred_cls_i):
+                for _j, pred_cls_i_j in enumerate(pred_cls_i):
                     conf_cls_i_j = self._score_function.get_set(
                         pred_cls=pred_cls_i_j,
                         quantile=self.lambda_classification,
@@ -716,30 +696,22 @@ class ODConformalizer(Conformalizer):
 
     """
 
-    MULTIPLE_TESTING_CORRECTIONS = ["bonferroni"]
-    BACKENDS = ["auto"]
-    GUARANTEE_LEVELS = ["image", "object"]
-    MATCHINGS = ["hausdorff", "iou", "giou", "lac", "mix"]
+    MULTIPLE_TESTING_CORRECTIONS = ("bonferroni",)
+    BACKENDS = ("auto",)
+    GUARANTEE_LEVELS = ("image", "object")
+    MATCHINGS = ("hausdorff", "iou", "giou", "lac", "mix")
 
     def __init__(
         self,
         backend: str = "auto",
         guarantee_level: str = "image",
         matching_function: str = "hausdorff",
-        confidence_threshold: Optional[float] = None,
-        multiple_testing_correction: Optional[str] = None,
-        confidence_method: Union[ConfidenceConformalizer, str, None] = None,
-        localization_method: Union[
-            LocalizationConformalizer,
-            str,
-            None,
-        ] = None,
+        confidence_threshold: float | None = None,
+        multiple_testing_correction: str | None = None,
+        confidence_method: ConfidenceConformalizer | str | None = None,
+        localization_method: LocalizationConformalizer | str | None = None,
         localization_prediction_set: str = "additive",  # Fix where we type check
-        classification_method: Union[
-            ClassificationConformalizer,
-            str,
-            None,
-        ] = None,
+        classification_method: ClassificationConformalizer | str | None = None,
         classification_prediction_set: str = "lac",  # Fix where we type check
         optimizer="binary_search",
         device="cpu",
@@ -784,10 +756,7 @@ class ODConformalizer(Conformalizer):
                 "No multiple_testing_correction provided, assuming no correction is needed. The explicit list of alphas is expected for calibration.",
             )
             self.multiple_testing_correction = multiple_testing_correction
-        elif (
-            multiple_testing_correction
-            not in self.MULTIPLE_TESTING_CORRECTIONS
-        ):
+        elif multiple_testing_correction not in self.MULTIPLE_TESTING_CORRECTIONS:
             raise ValueError(
                 f"multiple_testing_correction {multiple_testing_correction} not accepted, must be one of {self.MULTIPLE_TESTING_CORRECTIONS}",
             )
@@ -823,9 +792,7 @@ class ODConformalizer(Conformalizer):
         elif isinstance(localization_method, LocalizationConformalizer):
             self.localization_conformalizer = localization_method
             self.localization_method = localization_method.loss_name
-            self.localization_prediction_set = (
-                localization_method.prediction_set
-            )
+            self.localization_prediction_set = localization_method.prediction_set
         else:
             self.localization_conformalizer = None
             self.localization_method = None
@@ -846,9 +813,7 @@ class ODConformalizer(Conformalizer):
         elif isinstance(classification_method, ODClassificationConformalizer):
             self.classification_conformalizer = classification_method
             self.classification_method = classification_method.method
-            self.classification_prediction_set = (
-                classification_method.prediction_set
-            )
+            self.classification_prediction_set = classification_method.prediction_set
         else:
             self.classification_conformalizer = None
             self.classification_method = None
@@ -900,10 +865,10 @@ class ODConformalizer(Conformalizer):
     def calibrate(
         self,
         predictions: ODPredictions,
-        global_alpha: Optional[float] = None,
-        alpha_confidence: Optional[float] = None,
-        alpha_localization: Optional[float] = None,
-        alpha_classification: Optional[float] = None,
+        global_alpha: float | None = None,
+        alpha_confidence: float | None = None,
+        alpha_localization: float | None = None,
+        alpha_classification: float | None = None,
         verbose: bool = True,
     ) -> ODParameters:
         """Calibrates the conformalizers and returns the calibration results.
@@ -949,24 +914,9 @@ class ODConformalizer(Conformalizer):
                     "No multiple_testing_correction provided, expecting an explicit alpha for each conformalizer. 'global_alpha' should be 'None'.",
                 )
             if (
-                (
-                    alpha_confidence
-                    is None
-                    != self.confidence_conformalizer
-                    is None
-                )
-                or (
-                    alpha_localization
-                    is None
-                    != self.localization_conformalizer
-                    is None
-                )
-                or (
-                    alpha_classification
-                    is None
-                    != self.classification_conformalizer
-                    is None
-                )
+                (alpha_confidence is None != self.confidence_conformalizer is None)
+                or (alpha_localization is None != self.localization_conformalizer is None)
+                or (alpha_classification is None != self.classification_conformalizer is None)
             ):
                 raise ValueError(
                     "No multiple_testing_correction provided, expecting an explicit alpha for each conformalizer. Explicity alphas should be set only if there's a corresponding conformalizer.",
@@ -1123,7 +1073,7 @@ class ODConformalizer(Conformalizer):
     def conformalize(
         self,
         predictions: ODPredictions,
-        parameters: Optional[ODParameters] = None,
+        parameters: ODParameters | None = None,
         verbose: bool = True,
     ) -> ODConformalizedPredictions:
         """Conformalize the given predictions.
@@ -1152,11 +1102,9 @@ class ODConformalizer(Conformalizer):
                 logger.info(
                     "The parameters have been computed on another set of predictions.",
                 )
-            parameters_unique_id = parameters.unique_id
         else:
             if verbose:
                 logger.info("Using last parameters for conformalization")
-            parameters_unique_id = self._last_parameters_id
 
         # TODO: later check this is correct behavior
         if self.confidence_conformalizer is not None:
@@ -1313,10 +1261,12 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
 
     """
 
-    ACCEPTED_LOSSES = {
-        "pixelwise": PixelWiseRecallLoss,
-        "boxwise": BoxWiseRecallLoss,
-    }
+    ACCEPTED_LOSSES = MappingProxyType(
+        {
+            "pixelwise": PixelWiseRecallLoss,
+            "boxwise": BoxWiseRecallLoss,
+        }
+    )
 
     def __init__(
         self,
@@ -1361,19 +1311,13 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
 
         def risk_function(*lbd):
             lbd_loc, lbd_obj = lbd
-            pred_boxes_filtered = list(
-                [
-                    (
-                        x[y >= 1 - lbd_obj]
-                        if len(x[y >= 1 - lbd_obj]) > 0
-                        else x[None, y.argmax()]
-                    )
-                    for x, y in zip(
-                        predictions.pred_boxes,
-                        predictions.confidence,
-                    )
-                ],
-            )
+            pred_boxes_filtered = [
+                (x[y >= 1 - lbd_obj] if len(x[y >= 1 - lbd_obj]) > 0 else x[None, y.argmax()])
+                for x, y in zip(
+                    predictions.pred_boxes,
+                    predictions.confidence,
+                )
+            ]
             conf_boxes = apply_margins(
                 pred_boxes_filtered,
                 [lbd_loc, lbd_loc, lbd_loc, lbd_loc],
@@ -1415,7 +1359,7 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
         predictions: ODPredictions,
         alpha: float = 0.1,
         steps: int = 13,
-        bounds: list = [(0, 500), (0.0, 1.0)],
+        bounds: list | None = None,
         verbose: bool = True,
     ):
         """Calibrates the conformalizer using the given predictions.
@@ -1437,6 +1381,8 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
             ValueError: If the conformalizer has already been calibrated.
 
         """
+        if bounds is None:
+            bounds = [(0, 500), (0.0, 1.0)]
         if self.lbd is not None:
             logger.info("Replacing previously computed lambda")
         risk_function = self._get_risk_function(
@@ -1523,8 +1469,7 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
         ):
             cov = (
                 1
-                if len(true_boxes)
-                <= (confidence >= predictions.confidence_threshold).sum()
+                if len(true_boxes) <= (confidence >= predictions.confidence_threshold).sum()
                 else 0
             )
             set_size = (confidence >= predictions.confidence_threshold).sum()
@@ -1548,16 +1493,14 @@ class AsymptoticLocalizationObjectnessConformalizer(Conformalizer):
 
         conf_boxes = conf_boxes
         true_boxes = predictions.true_boxes
-        conf_boxes = list(
-            [
-                (
-                    x[y >= predictions.confidence_threshold]
-                    if len(x[y >= predictions.confidence_threshold]) > 0
-                    else x[None, y.argmax()]
-                )
-                for x, y in zip(conf_boxes, predictions.confidences)
-            ],
-        )
+        conf_boxes = [
+            (
+                x[y >= predictions.confidence_threshold]
+                if len(x[y >= predictions.confidence_threshold]) > 0
+                else x[None, y.argmax()]
+            )
+            for x, y in zip(conf_boxes, predictions.confidences)
+        ]
         set_size_loc = compute_set_size(conf_boxes)
         risk = compute_risk_object_level(
             conf_boxes,
