@@ -7,7 +7,7 @@ from cods.classif.data import ClassificationPredictions
 def plot_predictions(
     idxs: list,
     preds: ClassificationPredictions,
-    conf_cls: list = None,
+    conf_cls: list | None = None,
 ):
     if isinstance(idxs, int):
         idxs = [idxs]
@@ -17,7 +17,7 @@ def plot_predictions(
         )
 
     n = len(idxs)
-    fig, axs = plt.subplots(int(np.ceil(n / 4)), min(4, n), figsize=(12, 12))
+    _, axs = plt.subplots(int(np.ceil(n / 4)), min(4, n), figsize=(12, 12))
     for i, idx_pred in enumerate(idxs):
         curr_ax = axs[i // 4, i % 4] if n > 4 else (axs[i] if n > 1 else axs)
         image_path = preds.image_paths[idx_pred]
@@ -32,11 +32,9 @@ def plot_predictions(
                 f"True: {true_cls_name},\n Pred: {pred_cls_name}",
             )
         else:
-            conf_cls_name = list(
-                [preds.idx_to_cls[x.item()] for x in conf_cls[idx_pred]],
-            )
+            conf_cls_name = [preds.idx_to_cls[x.item()] for x in conf_cls[idx_pred]]
             curr_ax.set_title(
-                f"True: {true_cls_name},\n Pred: {pred_cls_name},\n Conf: {conf_cls_name if len(conf_cls_name) == 1 else 'len='+str(len(conf_cls_name))}",
+                f"True: {true_cls_name},\n Pred: {pred_cls_name},\n Conf: {conf_cls_name if len(conf_cls_name) == 1 else 'len=' + str(len(conf_cls_name))}",
             )
             curr_ax.axis("off")
     plt.show()
