@@ -56,13 +56,7 @@ def bayesod(
         tmp_cluster = torch.where(row).cpu().numpy().tolist()
 
         # TODO(leo) @luca : there's absolutely a better way to do that
-        cluster = list(
-            [
-                element
-                for element in tmp_cluster
-                if element not in already_used
-            ],
-        )
+        cluster = [element for element in tmp_cluster if element not in already_used]
 
         if len(cluster) > 0:
             clusters.append(cluster)
@@ -94,7 +88,7 @@ def bayesod(
         # merge the softmax probabilities in a weighted way
         # such that it is guaranteed to still be probability distribution
         new_cls = torch.zeros_like(pred_cls[0])
-        for i, c in enumerate(cluster):
+        for _, c in enumerate(cluster):
             new_cls += pred_cls[c] * confidences[c]
         new_cls /= confidences[cluster].sum()
         assert (new_cls.sum().item() - 1) < 1e-8
@@ -113,14 +107,10 @@ def bayesod(
 
 def filter_preds(preds, confidence_threshold=0.001):
     filters = [
-        conf > confidence_threshold
-        if (conf > confidence_threshold).any()
-        else conf.argmin(0)[None]
+        conf > confidence_threshold if (conf > confidence_threshold).any() else conf.argmin(0)[None]
         for conf in preds.confidences
     ]
     preds.pred_boxes = [pbs[f] for pbs, f in zip(preds.pred_boxes, filters)]
     preds.pred_cls = [pcs[f] for pcs, f in zip(preds.pred_cls, filters)]
-    preds.confidences = [
-        conf[f] for conf, f in zip(preds.confidences, filters)
-    ]
+    preds.confidences = [conf[f] for conf, f in zip(preds.confidences, filters)]
     return preds
