@@ -1,5 +1,5 @@
+from collections.abc import Callable
 from logging import getLogger
-from typing import List
 
 import numpy as np
 import torch
@@ -49,7 +49,6 @@ class FirstStepMonotonizingOptimizer(Optimizer):
         alpha: float,
         device: str,
         B: float = 1,
-        bounds: List[float] = [0, 1],  # deprecated
         init_lambda: float = 1,
         verbose: bool = False,
     ):
@@ -61,22 +60,16 @@ class FirstStepMonotonizingOptimizer(Optimizer):
         confidences = predictions.confidences
 
         stacked_confidences = torch.concatenate(
-            [x for x in confidences],  # .squeeze()
+            confidences,
         )
         confidence_image_idx = torch.concatenate(
-            [
-                torch.ones_like(x, dtype=int) * i
-                for i, x in enumerate(confidences)
-            ],
+            [torch.ones_like(x, dtype=int) * i for i, x in enumerate(confidences)],
         )
 
         sorted_stacked_confidences, indices = torch.sort(stacked_confidences)
+        sorted_stacked_confidences[:-1] = sorted_stacked_confidences[1:].clone()
+        sorted_stacked_confidences[-1] = 1.0  # last one is always 1.0
         sorted_confidence_image_indices = confidence_image_idx[indices]
-        sorted_confidence_image_indices[1:] = sorted_confidence_image_indices[
-            0:-1
-        ].clone()
-        # We let the first be, it should occur no change anyways ?
-        # sorted_confidence_image_indices[0] = ???
 
         lambda_conf = init_lambda
 
@@ -95,7 +88,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
 
         # TODO(leo):parallelize?
         # Step 1: Compute the risk
-        for i in tqdm(range(len(predictions))):
+        for i in tqdm(range(len(predictions)), disable=not verbose):
             true_boxes_i = true_boxes[i]
             pred_boxes_i = pred_boxes[i]
             confidences_i = confidences[i]
@@ -106,11 +99,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
             matching_i = predictions.matching[i]
 
             pred_boxes_i = pred_boxes_i[confidences_i >= 1 - lambda_conf]
-            pred_cls_i = [
-                x
-                for x, c in zip(pred_cls_i, confidences_i)
-                if c >= 1 - lambda_conf
-            ]
+            pred_cls_i = [x for x, c in zip(pred_cls_i, confidences_i) if c >= 1 - lambda_conf]
             pred_cls_i = (
                 torch.stack(pred_cls_i)
                 if len(pred_cls_i) > 0
@@ -138,19 +127,14 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                 if len(tmp_matched_boxes_i) > 0
                 else torch.tensor([]).float().to(device)
             )
-            # print(matched_pred_boxes_i.shape)
-            matched_pred_cls_i = list(
-                [
-                    (
-                        torch.stack([pred_cls_i[m] for m in matching_i[j]])[
-                            0
-                        ]  # TODO zero here ?
-                        if len(matching_i[j]) > 0
-                        else torch.tensor([]).float().to(device)
-                    )
-                    for j in range(len(true_boxes_i))
-                ],
-            )
+            matched_pred_cls_i = [
+                (
+                    torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]  # TODO zero here ?
+                    if len(matching_i[j]) > 0
+                    else torch.tensor([]).float().to(device)
+                )
+                for j in range(len(true_boxes_i))
+            ]
             margin = np.concatenate((image_shape, image_shape))
             matched_conf_boxes_i = apply_margins(
                 [matched_pred_boxes_i],
@@ -163,9 +147,6 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                 torch.arange(n_classes)[None, ...].to(device)
                 for _ in range(len(matched_pred_cls_i))
             ]
-
-            # if matched_conf_boxes_i.size() == 0:
-            #     matched_conf_boxes_i = torch.tensor([]).float().to(device)
 
             localization_loss_i = localization_loss(
                 true_boxes_i,
@@ -184,10 +165,6 @@ class FirstStepMonotonizingOptimizer(Optimizer):
             localization_losses.append(localization_loss_i)
             classification_losses.append(classification_loss_i)
 
-        # confidence_risk = torch.mean(torch.stack(confidence_losses))
-        # localization_risk = torch.mean(torch.stack(localization_losses))
-        # classification_risk = torch.mean(torch.stack(classification_losses))
-
         # TODO(leo): image level
         confidence_risk = self._correct_risk(
             confidence_losses,
@@ -205,27 +182,32 @@ class FirstStepMonotonizingOptimizer(Optimizer):
             B,
         )
 
+        # ------- LOGGING -------
+        _log_raw_confidence_losses = confidence_losses.copy()
+        _log_raw_localization_losses = localization_losses.copy()
+        _log_raw_classification_losses = classification_losses.copy()
+
         max_risk = torch.max(
             torch.stack(
                 [confidence_risk, localization_risk, classification_risk],
             ),
         )
-        print(f"First risk: {max_risk.detach().cpu().numpy()}")
+        logger.info(f"First risk: {max_risk.detach().cpu().numpy()}")
         if max_risk.detach().cpu().numpy() > alpha:
             # Debug: all three risks to see why there isn't any solution
             logger.debug(f"Confidence risk: {confidence_risk}")
             logger.debug(f"Localization risk: {localization_risk}")
             logger.debug(f"Classification risk: {classification_risk}")
             logger.debug(f"Max risk: {max_risk} > {alpha}. No solution found.")
-            raise ValueError(
+            logger.warning(
                 "There does not exist any solution satisfying the constraints.",
             )
+            return 1.0
         logger.debug(
             f"Risk after 1st epoch is {max_risk.detach().cpu().numpy()} < {alpha}",
         )
 
         previous_lbd = lambda_conf
-        previous_risk = max_risk
 
         pbar = tqdm(
             list(
@@ -234,6 +216,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                     sorted_stacked_confidences,
                 ),
             ),
+            disable=not verbose,
         )
 
         self.all_risks_raw = [max_risk.detach().cpu().numpy()]
@@ -249,8 +232,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
         # Step 2: Update one loss at a time
         for image_id, conf_score in pbar:
             previous_lbd = lambda_conf
-            previous_risk = max_risk
-            lambda_conf = 1 - conf_score.cpu().numpy()  # - 1e-7  # Test
+            lambda_conf = 1 - conf_score.cpu().numpy().item()
             if lambda_conf > init_lambda:
                 continue
 
@@ -263,11 +245,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
             image_shape = image_shapes[i]
 
             pred_boxes_i = pred_boxes_i[confidences_i >= 1 - lambda_conf]
-            pred_cls_i = [
-                x
-                for x, c in zip(pred_cls_i, confidences_i)
-                if c >= 1 - lambda_conf
-            ]
+            pred_cls_i = [x for x, c in zip(pred_cls_i, confidences_i) if c >= 1 - lambda_conf]
             pred_cls_i = (
                 torch.stack(pred_cls_i)
                 if len(pred_cls_i) > 0
@@ -291,19 +269,6 @@ class FirstStepMonotonizingOptimizer(Optimizer):
 
             predictions.matching[i] = matching_i
 
-            # if i in [14, 74, 199, 213, 225, 234]:
-            #     print("--------------------------------------------------")
-            #     print(f"Image {i}")
-            #     print(
-            #         f"Confidence Loss: {confidence_loss_i.detach().cpu().numpy()}"
-            #     )  # .tolist:.4f}")
-            #     print(f"Number of ground truths: {len(true_boxes_i)}")
-            #     print(f"Number of predictions: {len(pred_boxes_i)}")
-            #     print(f"Confidences ({confidences_i.dtype}):")
-            #     for c in confidences_i:
-            #         print(f"\t{c}")
-            #     print(f"Lambda and confidence: {lambda_conf} and {conf_score}")
-
             tmp_matched_boxes_i = [
                 (
                     torch.stack([pred_boxes_i[m] for m in matching_i[j]])[0]
@@ -317,16 +282,14 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                 if len(tmp_matched_boxes_i) > 0
                 else torch.tensor([]).float().to(device)
             )
-            matched_pred_cls_i = list(
-                [
-                    (
-                        torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
-                        if len(matching_i[j]) > 0
-                        else torch.tensor([]).float().to(device)
-                    )
-                    for j in range(len(true_boxes_i))
-                ],
-            )
+            matched_pred_cls_i = [
+                (
+                    torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
+                    if len(matching_i[j]) > 0
+                    else torch.tensor([]).float().to(device)
+                )
+                for j in range(len(true_boxes_i))
+            ]
 
             margin = np.concatenate((image_shape, image_shape))
             matched_conf_boxes_i = apply_margins(
@@ -340,12 +303,7 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                 torch.arange(n_classes)[None, ...].to(device)
                 for _ in range(len(matched_pred_cls_i))
             ]
-            # print(
-            #     true_boxes_i.shape,
-            #     true_cls_i.shape,
-            #     matched_conf_boxes_i[0].shape,
-            #     matched_conf_cls_i[0].shape,
-            # )
+
             localization_loss_i = localization_loss(
                 true_boxes_i,
                 true_cls_i,
@@ -359,14 +317,22 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                 matched_conf_cls_i,
             )
 
-            # if confidence_loss_i.detach().cpu().numpy()[0] != confidence_losses[i].detach().cpu().numpy()[0]:
-            #     print(f"Confidence Loss: {confidence_loss_i.detach().cpu().numpy()}")#.tolist:.4f}")
-            #     print(f"Localization Loss: {localization_loss_i.detach().cpu().numpy()}")#:.4f}")
-            #     print(f"Classification Loss: {classification_loss_i.detach().cpu().numpy()}")#:.4f}")
+            _log_raw_confidence_losses[i] = confidence_loss_i.detach().clone()
+            _log_raw_localization_losses[i] = localization_loss_i.detach().clone()
+            _log_raw_classification_losses[i] = classification_loss_i.detach().clone()
 
-            confidence_losses[i] = confidence_loss_i.detach().clone()
-            localization_losses[i] = localization_loss_i.detach().clone()
-            classification_losses[i] = classification_loss_i.detach().clone()
+            confidence_losses[i] = max(
+                confidence_loss_i.detach().clone(),
+                confidence_losses[i].clone(),
+            )
+            localization_losses[i] = max(
+                localization_loss_i.detach().clone(),
+                localization_losses[i].clone(),
+            )
+            classification_losses[i] = max(
+                classification_loss_i.detach().clone(),
+                classification_losses[i].clone(),
+            )
 
             confidence_risk = self._correct_risk(
                 confidence_losses,
@@ -389,47 +355,41 @@ class FirstStepMonotonizingOptimizer(Optimizer):
                     [confidence_risk, localization_risk, classification_risk],
                 ),
             )
-            # idddd = torch.argmax(
-            #     torch.stack(
-            #         [confidence_risk, localization_risk, classification_risk]
-            #     )
-            # )
-            # logger.info(f"maximizing risk : {idddd} where 0 is confidence, 1 is localization and 2 is classification")
-
+            _log_raw_confidence_risk = self._correct_risk(
+                _log_raw_confidence_losses,
+                len(predictions),
+                B,
+            )
+            _log_raw_localization_risk = self._correct_risk(
+                _log_raw_localization_losses,
+                len(predictions),
+                B,
+            )
+            _log_raw_classification_risk = self._correct_risk(
+                _log_raw_classification_losses,
+                len(predictions),
+                B,
+            )
+            _log_raw_max_risk = torch.max(
+                torch.stack(
+                    [
+                        _log_raw_confidence_risk,
+                        _log_raw_localization_risk,
+                        _log_raw_classification_risk,
+                    ],
+                ),
+            )
             self.all_lbds.append(lambda_conf)
-            # _tmp_max_risk =
-            self.all_risks_raw.append(max_risk.detach().cpu().numpy())
+            self.all_risks_raw.append(_log_raw_max_risk.detach().cpu().numpy())
             self.all_risks_raw_conf.append(
-                confidence_risk.detach().cpu().numpy(),
+                _log_raw_confidence_risk.detach().cpu().numpy(),
             )
             self.all_risks_raw_loc.append(
-                localization_risk.detach().cpu().numpy(),
+                _log_raw_localization_risk.detach().cpu().numpy(),
             )
             self.all_risks_raw_cls.append(
-                classification_risk.detach().cpu().numpy(),
+                _log_raw_classification_risk.detach().cpu().numpy(),
             )
-
-            # Monotonization
-            if max_risk < previous_risk:
-                max_risk = previous_risk
-                confidence_risk = np.max(
-                    [
-                        self.all_risks_mon_conf[-1],
-                        confidence_risk.detach().cpu().numpy(),
-                    ],
-                )
-                localization_risk = np.max(
-                    [
-                        self.all_risks_mon_loc[-1],
-                        localization_risk.detach().cpu().numpy(),
-                    ],
-                )
-                classification_risk = np.max(
-                    [
-                        self.all_risks_mon_cls[-1],
-                        classification_risk.detach().cpu().numpy(),
-                    ],
-                )
 
             self.all_risks_mon.append(max_risk.detach().cpu().numpy())
             self.all_risks_mon_conf.append(
@@ -492,13 +452,9 @@ class FirstStepMonotonizingOptimizer(Optimizer):
 
                     matching_i = predictions.matching[i]
 
-                    pred_boxes_i = pred_boxes_i[
-                        confidences_i >= 1 - previous_lbd
-                    ]
+                    pred_boxes_i = pred_boxes_i[confidences_i >= 1 - previous_lbd]
                     pred_cls_i = [
-                        x
-                        for x, c in zip(pred_cls_i, confidences_i)
-                        if c >= 1 - previous_lbd
+                        x for x, c in zip(pred_cls_i, confidences_i) if c >= 1 - previous_lbd
                     ]
                     confidence_loss_i = confidence_loss(
                         true_boxes_i,
@@ -575,29 +531,27 @@ class SecondStepMonotonizingOptimizer(Optimizer):
         build_predictions,
         matching_function,
     ):
+        all_risks_raw = []
+        all_risks_mon = []
+        all_lbds_cnf = []
+        all_lbds_loc = []
+
         true_boxes = predictions.true_boxes
         pred_boxes = predictions.pred_boxes
         true_cls = predictions.true_cls
         pred_cls = predictions.pred_cls
-        image_shapes = predictions.image_shapes
         confidences = predictions.confidences
         device = predictions.true_boxes[0].device
 
         stacked_confidences = torch.concatenate(confidences)
         confidence_image_idx = torch.concatenate(
-            [
-                torch.ones_like(x, dtype=torch.int) * i
-                for i, x in enumerate(confidences)
-            ],
+            [torch.ones_like(x, dtype=torch.int) * i for i, x in enumerate(confidences)],
         )
 
         sorted_stacked_confidences, indices = torch.sort(stacked_confidences)
+        sorted_stacked_confidences[:-1] = sorted_stacked_confidences[1:].clone()
+        sorted_stacked_confidences[-1] = 1.0  # last one is always 1.0
         sorted_confidence_image_indices = confidence_image_idx[indices]
-        sorted_confidence_image_indices[1:] = sorted_confidence_image_indices[
-            0:-1
-        ].clone()
-        # We let the first be, it should occur no change anyways ?
-        # sorted_confidence_image_indices[0] = ???
 
         lambda_conf = 1
 
@@ -610,14 +564,23 @@ class SecondStepMonotonizingOptimizer(Optimizer):
 
         losses = []
 
-        # TODO(leo):parallelize?
+        # TODO(leo): parallelize?
         # Step 1: Compute the risk
         for i in range(len(predictions)):
             true_boxes_i = true_boxes[i]
             pred_boxes_i = pred_boxes[i]
             true_cls_i = true_cls[i]
             pred_cls_i = pred_cls[i]
+            confidences_i = confidences[i]
             matching_i = predictions.matching[i]
+
+            pred_boxes_i = pred_boxes_i[confidences_i >= 1 - lambda_conf]
+            pred_cls_i = [x for x, c in zip(pred_cls_i, confidences_i) if c >= 1 - lambda_conf]
+            pred_cls_i = (
+                torch.stack(pred_cls_i)
+                if len(pred_cls_i) > 0
+                else torch.tensor([]).float().to(device)
+            )
 
             tmp_matched_boxes_i = [
                 (
@@ -632,21 +595,14 @@ class SecondStepMonotonizingOptimizer(Optimizer):
                 if len(tmp_matched_boxes_i) > 0
                 else torch.tensor([]).float().to(device)
             )
-            # print(matched_pred_boxes_i.shape)
-            matched_pred_cls_i = list(
-                [
-                    (
-                        torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
-                        if len(matching_i[j]) > 0
-                        else torch.tensor([]).float().to(device)
-                    )
-                    for j in range(len(true_boxes_i))
-                ],
-            )
-            # if len(matched_pred_boxes_i.shape)==1:
-            #     matched_pred_boxes_i = matched_pred_boxes_i[None,...]
-            # if len(matched_pred_cls_i.shape)==1:
-            #     matched_pred_cls_i = matched_pred_cls_i[None,...]
+            matched_pred_cls_i = [
+                (
+                    torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
+                    if len(matching_i[j]) > 0
+                    else torch.tensor([]).float().to(device)
+                )
+                for j in range(len(true_boxes_i))
+            ]
 
             matched_conf_boxes_i, matched_conf_cls_i = build_predictions(
                 matched_pred_boxes_i,
@@ -663,23 +619,28 @@ class SecondStepMonotonizingOptimizer(Optimizer):
 
             losses.append(loss_i)
 
-        risk = torch.mean(torch.stack(losses))
-
-        n_losses = len(losses)
+        # ------- Start Logging -------
+        _log_risk = torch.mean(torch.stack(losses))
+        _log_losses = losses.copy()
+        all_risks_raw.append(_log_risk.detach().cpu().numpy())
+        all_risks_mon.append(_log_risk.detach().cpu().numpy())
+        all_lbds_cnf.append(lambda_conf)
+        all_lbds_loc.append(lbd)
+        # ------- End Logging -------
 
         # Step 2: Update one loss at a time
         for image_id, conf_score in zip(
             sorted_confidence_image_indices,
             sorted_stacked_confidences,
         ):
-            previous_risk = risk
-            lambda_conf = 1 - conf_score.cpu().numpy()
+            lambda_conf = 1 - conf_score.cpu().numpy().item()
 
             i = image_id
             true_boxes_i = true_boxes[i]
             pred_boxes_i = pred_boxes[i]
             true_cls_i = true_cls[i]
             pred_cls_i = pred_cls[i]
+            confidences_i = confidences[i]
 
             matching_i = match_predictions_to_true_boxes(
                 predictions,
@@ -691,6 +652,14 @@ class SecondStepMonotonizingOptimizer(Optimizer):
 
             predictions.matching[i] = matching_i
 
+            pred_boxes_i = pred_boxes_i[confidences_i >= 1 - lambda_conf]
+            pred_cls_i = [x for x, c in zip(pred_cls_i, confidences_i) if c >= 1 - lambda_conf]
+            pred_cls_i = (
+                torch.stack(pred_cls_i)
+                if len(pred_cls_i) > 0
+                else torch.tensor([]).float().to(device)
+            )
+
             # TODO: currently only support matching to a single box
             tmp_matched_boxes_i = [
                 (
@@ -700,26 +669,19 @@ class SecondStepMonotonizingOptimizer(Optimizer):
                 )
                 for j in range(len(true_boxes_i))
             ]
-            # print([x.shape for x in tmp_matched_boxes_i])
             matched_pred_boxes_i = (
                 torch.stack(tmp_matched_boxes_i)
                 if len(tmp_matched_boxes_i) > 0
                 else torch.tensor([]).float().to(device)
             )
-            matched_pred_cls_i = list(
-                [
-                    (
-                        torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
-                        if len(matching_i[j]) > 0
-                        else torch.tensor([]).float().to(device)
-                    )
-                    for j in range(len(true_boxes_i))
-                ],
-            )
-
-            # print(matched_pred_boxes_i.shape)
-            # print(matched_pred_boxes_i)
-
+            matched_pred_cls_i = [
+                (
+                    torch.stack([pred_cls_i[m] for m in matching_i[j]])[0]
+                    if len(matching_i[j]) > 0
+                    else torch.tensor([]).float().to(device)
+                )
+                for j in range(len(true_boxes_i))
+            ]
             matched_conf_boxes_i, matched_conf_cls_i = build_predictions(
                 matched_pred_boxes_i,
                 matched_pred_cls_i,
@@ -733,66 +695,86 @@ class SecondStepMonotonizingOptimizer(Optimizer):
                 matched_conf_cls_i,
             )
 
-            old_loss_i = losses[i]
+            _log_losses[i] = loss_i.clone()
+            old_loss_i = losses[i].clone()
+            loss_i = max(old_loss_i, loss_i)
             losses[i] = loss_i
 
-            # risk = torch.mean(torch.stack(losses))
-            # Faster
-            risk = risk + (loss_i - old_loss_i) / n_losses
-
-            self.all_risks_raw.append(risk.detach().cpu().numpy())
-
-            risk = max(risk, previous_risk)
-
-            self.all_risks_mon.append(risk.detach().cpu().numpy())
-
-            self.all_lbds_cnf.append(lambda_conf)
-            self.all_lbds_loc.append(lbd)
+            # ------- Start Logging -------
+            _log_mon_risk = torch.mean(torch.stack(losses))
+            _log_raw_risk = torch.mean(torch.stack(_log_losses))
+            all_risks_raw.append(_log_raw_risk.detach().cpu().numpy())
+            all_risks_mon.append(_log_mon_risk.detach().cpu().numpy())
+            all_lbds_cnf.append(lambda_conf)
+            all_lbds_loc.append(lbd)
+            # ------- End Logging -------
 
             # Stopping condition: when we reached desired lbd_conf
             if final_lbd_conf >= lambda_conf:
-                return risk
+                return (
+                    torch.mean(torch.stack(losses)).detach().cpu().numpy(),
+                    all_risks_raw,
+                    all_risks_mon,
+                    all_lbds_cnf,
+                    all_lbds_loc,
+                )
+        return None, all_risks_raw, all_risks_mon, all_lbds_cnf, all_lbds_loc
 
     def optimize(
         self,
         predictions: ODPredictions,
-        build_predictions,
+        build_predictions: Callable,
         loss: ODLoss,
-        matching_function,
+        matching_function: str,
         alpha: float,
         device: str,
+        overload_confidence_threshold: float | None = None,
         B: float = 1,
-        bounds: List[float] = [0, 1],
-        steps=13,
-        epsilon=1e-10,
+        lower_bound: float = 0,
+        upper_bound: float = 1,
+        steps: int = 13,
+        epsilon: float = 1e-10,
+        *,
         verbose: bool = False,
     ):
+        if overload_confidence_threshold is None:
+            if predictions.confidence_threshold is None:
+                raise ValueError(
+                    "confidence_threshold must be set in the predictions or in the conformalizer",
+                )
+            confidence_threshold = predictions.confidence_threshold
+            if isinstance(confidence_threshold, torch.Tensor):
+                confidence_threshold = confidence_threshold.item()
+            logger.info(
+                f"Using predictions' confidence threshold: {confidence_threshold:.4f}",
+            )
+        else:
+            logger.info(
+                f"Using overload confidence threshold: {overload_confidence_threshold:.4f}",
+            )
+            confidence_threshold = overload_confidence_threshold
+
+        ##### FINISH MODIFYING THE LOGGING OUTPUT AND COLLECTION
+
         self.all_risks_raw = []
         self.all_risks_mon = []
         self.all_lbds_cnf = []
         self.all_lbds_loc = []
 
-        lambda_conf = 1 - predictions.confidence_threshold
+        lambda_conf = 1 - confidence_threshold
 
         # Step 0: Initialize the risk
 
-        match_predictions_to_true_boxes(
-            predictions,
-            distance_function=matching_function,
-            verbose=False,
-            overload_confidence_threshold=1 - lambda_conf,
-        )
-
-        left, right = bounds
-        lbd = (left + right) / 2
+        left, right = lower_bound, upper_bound
 
         good_lbds = []
 
         pbar = tqdm(range(steps), disable=not verbose)
 
-        for step in pbar:
+        for _ in pbar:
+            lbd = (left + right) / 2
             # Evaluating the risk in this lbd, requires to remonotonize the loss in this lbd_loc/cls wrt the lbd_cnf
-            risk = self.evaluate_risk(
+            risk, all_risks_raw, all_risks_mon, all_lbds_cnf, all_lbds_loc = self.evaluate_risk(
                 lbd,
                 loss,
                 lambda_conf,
@@ -800,25 +782,26 @@ class SecondStepMonotonizingOptimizer(Optimizer):
                 build_predictions,
                 matching_function,
             )
+            self.all_risks_raw.append(all_risks_raw)
+            self.all_risks_mon.append(all_risks_mon)
+            self.all_lbds_cnf.append(all_lbds_cnf)
+            self.all_lbds_loc.append(all_lbds_loc)
 
             corrected_risk = self._correct_risk(risk, len(predictions), B)
 
-            corrected_risk = corrected_risk.detach().cpu().numpy().item()
+            corrected_risk = corrected_risk.item()  # detach().cpu().item()
 
             pbar.set_description(
-                f"[{left:.2f}, {right:.2f}] -> λ={lbd}. Corrected Risk = {corrected_risk:.2f}",
+                f"[{left:.2f}, {right:.2f}] -> λ={lbd}. Corrected Risk = {corrected_risk:.3f}",
             )
 
-            if risk <= alpha:
-                good_lbds.append(lbd)
-                if risk >= alpha - epsilon:
-                    break
-
-            if corrected_risk > alpha:
-                left = lbd
-            else:
+            if corrected_risk <= alpha:
                 right = lbd
-            lbd = (left + right) / 2
+                good_lbds.append(lbd)
+                if corrected_risk >= alpha - epsilon:
+                    break
+            else:  # corrected_risk > alpha
+                left = lbd
 
         if len(good_lbds) == 0:
             raise ValueError("No good lambda found")
