@@ -12,7 +12,7 @@ def plot_preds(
     predictions: ODPredictions,
     conformalized_predictions: ODConformalizedPredictions = None,
     confidence_threshold=None,
-    idx_to_label: dict = None,
+    idx_to_label: dict | None = None,
     save_as=None,
 ):
     """Plot the predictions of an object detection model.
@@ -40,10 +40,7 @@ def plot_preds(
         conf_boxes = conformalized_predictions.conf_boxes[idx]
         conf_cls = conformalized_predictions.conf_cls[idx]
 
-    if (
-        confidence_threshold is None
-        and predictions.confidence_threshold is not None
-    ):
+    if confidence_threshold is None and predictions.confidence_threshold is not None:
         confidence_threshold = predictions.confidence_threshold
         print("Using confidence threshold from preds")
     else:
@@ -104,7 +101,7 @@ def plot_preds(
                     y1,
                     text,
                     fontsize=15,
-                    bbox=dict(facecolor=color, alpha=0.5),
+                    bbox={"facecolor": color, "alpha": 0.5},
                 )
             else:
                 # Print nb of labels
@@ -114,7 +111,7 @@ def plot_preds(
                     y1,
                     text,
                     fontsize=15,
-                    bbox=dict(facecolor=color, alpha=0.5),
+                    bbox={"facecolor": color, "alpha": 0.5},
                 )
         elif isinstance(proba, int) or len(proba.shape) == 0:
             if isinstance(proba, torch.Tensor):
@@ -128,7 +125,7 @@ def plot_preds(
                 y2,
                 text,
                 fontsize=15,
-                bbox=dict(facecolor=color, alpha=0.5),
+                bbox={"facecolor": color, "alpha": 0.5},
             )
         else:
             cl = proba.argmax().item()
@@ -142,7 +139,7 @@ def plot_preds(
                 y1,
                 text,
                 fontsize=15,
-                bbox=dict(facecolor=color, alpha=0.5),
+                bbox={"facecolor": color, "alpha": 0.5},
             )
 
     ax = plt.gca()
@@ -165,7 +162,7 @@ def plot_preds(
             "",
             xy=(true_box[0], true_box[1]),
             xytext=(matching_pred_box[0], matching_pred_box[1]),
-            arrowprops=dict(arrowstyle="->", lw=2, color="blue"),
+            arrowprops={"arrowstyle": "->", "lw": 2, "color": "blue"},
         )
 
     plt.axis("off")
@@ -178,7 +175,7 @@ def create_pdf_with_plots(
     predictions: ODPredictions,
     conformalized_predictions: ODConformalizedPredictions = None,
     confidence_threshold=None,
-    idx_to_label: dict = None,
+    idx_to_label: dict | None = None,
     output_pdf="output.pdf",
 ):
     """Create a PDF with plots for each image in the predictions.
@@ -194,10 +191,7 @@ def create_pdf_with_plots(
     """
     is_conformal = conformalized_predictions is not None
 
-    if (
-        confidence_threshold is None
-        and predictions.confidence_threshold is not None
-    ):
+    if confidence_threshold is None and predictions.confidence_threshold is not None:
         confidence_threshold = predictions.confidence_threshold
     elif confidence_threshold is None:
         raise ValueError("Confidence Threshold should be provided")
@@ -227,7 +221,15 @@ def create_pdf_with_plots(
             plt.figure(figsize=(14, 14))
             plt.imshow(image)
 
-            def draw_rect(ax, box, color, proba, conformal=False):
+            def draw_rect(
+                ax,
+                box,
+                color,
+                proba,
+                conformal=False,
+                image_width=image_width,
+                image_height=image_height,
+            ):
                 x1, y1, x2, y2 = box
                 x1 = max(0, x1)
                 y1 = max(0, y1)
@@ -259,7 +261,7 @@ def create_pdf_with_plots(
                             y1,
                             text,
                             fontsize=15,
-                            bbox=dict(facecolor=color, alpha=0.5),
+                            bbox={"facecolor": color, "alpha": 0.5},
                         )
                     else:
                         text = f"{len(proba)} labels"
@@ -268,15 +270,13 @@ def create_pdf_with_plots(
                             y1,
                             text,
                             fontsize=15,
-                            bbox=dict(facecolor=color, alpha=0.5),
+                            bbox={"facecolor": color, "alpha": 0.5},
                         )
                 elif isinstance(proba, int) or len(proba.shape) == 0:
                     if isinstance(proba, torch.Tensor):
                         proba = proba.item()
                     if idx_to_label is not None:
-                        text = (
-                            f"{idx_to_label[proba]}" if proba >= 0 else "conf"
-                        )
+                        text = f"{idx_to_label[proba]}" if proba >= 0 else "conf"
                     else:
                         text = f"{proba}" if proba >= 0 else "conf"
                     ax.text(
@@ -284,7 +284,7 @@ def create_pdf_with_plots(
                         y2,
                         text,
                         fontsize=15,
-                        bbox=dict(facecolor=color, alpha=0.5),
+                        bbox={"facecolor": color, "alpha": 0.5},
                     )
                 else:
                     cl = proba.argmax().item()
@@ -298,7 +298,7 @@ def create_pdf_with_plots(
                         y1,
                         text,
                         fontsize=15,
-                        bbox=dict(facecolor=color, alpha=0.5),
+                        bbox={"facecolor": color, "alpha": 0.5},
                     )
 
             ax = plt.gca()
@@ -324,7 +324,7 @@ def create_pdf_with_plots(
                         "",
                         xy=(true_box[0], true_box[1]),
                         xytext=(matching_pred_box[0], matching_pred_box[1]),
-                        arrowprops=dict(arrowstyle="->", lw=2, color="blue"),
+                        arrowprops={"arrowstyle": "->", "lw": 2, "color": "blue"},
                     )
                 except Exception as e:
                     print(e)
@@ -342,12 +342,9 @@ def plot_histograms_predictions(predictions: ODPredictions):
 
     list_true = [len(x) for x in predictions.true_boxes]
     list_pred = [len(x) for x in predictions.confidences]
-    list_pred_thresh = [
-        sum(x > predictions.confidence_threshold)
-        for x in predictions.confidence
-    ]
+    list_pred_thresh = [sum(x > predictions.confidence_threshold) for x in predictions.confidence]
 
-    fig, axs = plt.subplots(1, 3, figsize=(15, 5))
+    _, axs = plt.subplots(1, 3, figsize=(15, 5))
     axs[0].hist(list_true, bins=20)
     axs[0].set_title(f"True boxes, mean size = {np.mean(list_true):.2f}")
     axs[1].hist(list_pred, bins=20)
